docs(blueprint): minimize plasTeX package overrides (#400)
