+++ blueprint/src/_tnlean_patches.sty
@@ -1,0 +1,3 @@
+\NeedsTeXFormat{LaTeX2e}
+\ProvidesPackage{_tnlean_patches}[2026/04/18 No-op stub for plasTeX-only TNLean patches]
+\endinput
