+++ blueprint/src/web.tex
@@ -5,6 +5,7 @@
 \usepackage{natbib}
 \usepackage{hyperref}
 \usepackage[showmore, dep_graph]{blueprint}
+\usepackage{_tnlean_patches}
 
 \input{macros/common}
 \input{macros/web}
